refactor: remove unused components and files
