+++ client/src/js/app/store/modules/store.proposal.js
@@ -113,7 +113,7 @@ const proposalModule = {
       },
   },
   getters: {
-    currentProposal: state  => {
+    currentProposal: state => {
       // If we have no proposal set, check if there is one in storage
       // Should not need to do this now.... TODO - simplify this
       if (!state.proposal) {
