+++ client/src/js/modules/samples/components/ContainerGraphic.vue
