--- client/src/js/modules/samples/components/Containers.vue
@@ -1,1552 +1,0 @@
-<template>
-    <section class="tw-container tw-mx-auto">
-        <div class="tw-mb-4">
-            <h1 class="tw-text-xl">Plate Geometry</h1>
-            <label class="">Capacity</label>
-            <input v-model="container.capacity" class="tw-border tw-rounded tw-shadow tw-px-2" type="number"/>
-            <label>Rows</label>
-            <input v-model="container.rows" class="tw-border tw-rounded tw-shadow tw-px-2" type="number"/>
-        </div>
-        <div class="tw-mb-4">
-            <label class="">Drops X</label>
-            <input v-model="container.drops.x" class="tw-border tw-rounded tw-shadow tw-px-2" type="number"/>
-            <label class="">Drops Y</label>
-            <input v-model="container.drops.y" class="tw-border tw-rounded tw-shadow tw-px-2" type="number"/>
-            <label class="">Drop Width</label>
-            <input v-model="container.drops.w" class="tw-border tw-rounded tw-shadow tw-px-2" type="number" max="1" min="0"/>
-            <label class="">Drop Height</label>
-            <input v-model="container.drops.h" class="tw-border tw-rounded tw-shadow tw-px-2" type="number" max="1" min="0"/>
-        </div>
-        <div class="tw-mb-4">
-            <button class="tw-border tw-rounded tw-px-2 tw-py-1 tw-bg-blue-400 hover:tw-bg-blue-600 tw-text-white" @click="updatePlateView" type="submit">Update Plate Geometry </button>
-        </div>
-        <div class="tw-mb-4">
-            <h1 class="tw-text-xl">Scoring Color Schemes</h1>
-            <label>Color Scale</label>
-            <select v-model="colorScale" class="tw-border tw-rounded tw-shadow tw-px-2">
-                <option value="viridis">Viridis</option>
-                <option value="rgb">RGB</option>
-                <option value="threshold">Threshold</option>
-            </select>
-            <div v-if="colorScale == 'threshold'">
-                <label>Threshold Value: </label><input v-model="threshold" class="tw-align-bottom slider" type="range" min="0" max="100" id="myRange"/>
-                <div><p>Threshold: {{threshold}}%</p></div>
-            </div>
-        </div>
-
-        <p v-show="selectedItems.length">Selected Samples:</p>
-
-        <ul>
-            <li class="inline" v-for="(i,k) in selectedItems" :key="k">{{i}}, </li>
-        </ul>
-
-        <plate-view
-            :container="container"
-            :samples="samples.data"
-            :threshold="threshold/100"
-            :color-scale="colorScale"
-            :selected="selectedItems"
-            label="SCORE"
-            :key="plateKey"
-            @cell-clicked="onCellClicked">
-        </plate-view>
-        <puck-view
-            :samples="samples.data"
-            :threshold="threshold/100"
-            :color-scale="colorScale"
-            :selected="selectedItems"
-            @cell-clicked="onCellClicked">
-        </puck-view>
-    </section>
-</template>
-
-<script>
-import PlateView from './PlateView.vue'
-import PuckView from './PuckView.vue'
-
-export default {
-  name:"Containers",
-  components: {
-    'plate-view': PlateView,
-    'puck-view': PuckView,
-  },
-
-  methods: {
-    // Trick to rerender component if container geometry changes
-    updatePlateView: function() {
-      this.plateKey += 1;
-    },
-    onSave: function() {
-      alert("Saving Sample Group: " + this.sampleGroupName)
-    },
-    onCellClicked: function(location) {
-        console.log("Puck Cell location clicked event " + JSON.stringify(location))
-
-        let index = this.selectedItems.indexOf(location)
-
-        console.log(JSON.stringify(this.selectedItems))
-        console.log("Index for location " + location + " " + index)
-
-        if ( index < 0) {
-          console.log("Adding location " + location + " to selected array")
-          this.selectedItems.push(+location)
-        }
-        else this.selectedItems.splice(index, 1)
-    },
-  },
-  watch: {
-    threshold: function() {
-      this.updatePlateView()
-    },
-    colorScale: function() {
-      this.updatePlateView()
-     },
-  },
-  data: function() {
-    return {
-        selectedItems: [],
-        sampleGroupName: '',
-        plateKey: 0,
-        puckKey: 0,
-        container: {
-            capacity: 96*2,
-            columns: 12, // wells per row
-            drops: {x: 2, y: 1, h: 0.5, w: 1}, // Try x,y with a width factor
-            well: null, // or -1 to indicate we don't need
-            // Example 3 drop plate with well
-            // capacity: 96*3,
-            // columns: 12,
-            // drops: {x: 2, y: 2, h: 1, w: 1}, // Try x,y with a width factor
-            // well: 1
-        },
-        threshold: 50, // Above number show as green (x10 for range slider)
-        colorScale: 'rgb', // 'viridis', rgb', 'threshold'
-        // Sample data taken from a json response
-        samples: {
-        "total": 20,
-        "data": [
-            {
-                "BLSAMPLEID": "399094",
-                "CRYSTALID": "333570",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "1",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A1d1_A1d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 1.0,
-                "VALID": 0
-            },
-            {
-                "BLSAMPLEID": "499094",
-                "CRYSTALID": "333570",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "2",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A1d2_A1d2_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.9,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399095",
-                "CRYSTALID": "333571",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "3",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A2d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.7,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399096",
-                "CRYSTALID": "333572",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "5",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A3d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.5,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399097",
-                "CRYSTALID": "333573",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "7",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A4d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.4,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399098",
-                "CRYSTALID": "333574",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "9",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A5d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399099",
-                "CRYSTALID": "333575",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "11",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A6d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399100",
-                "CRYSTALID": "333576",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "13",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A7d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.3,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399101",
-                "CRYSTALID": "333577",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "15",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A8d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399102",
-                "CRYSTALID": "333578",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "17",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A9d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399103",
-                "CRYSTALID": "333579",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "19",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A10d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399104",
-                "CRYSTALID": "333580",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "21",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A11d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399105",
-                "CRYSTALID": "333581",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "23",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "A12d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399106",
-                "CRYSTALID": "333582",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "25",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "B1d1_A1d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399107",
-                "CRYSTALID": "333583",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "49",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "C1d1_A1d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399108",
-                "CRYSTALID": "333584",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "73",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "D1d1_A1d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399109",
-                "CRYSTALID": "333585",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "97",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "E1d1_A1d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399110",
-                "CRYSTALID": "333586",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "121",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "F1d1_A1d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399111",
-                "CRYSTALID": "333587",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "145",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "G1d1_A1d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            },
-            {
-                "BLSAMPLEID": "399112",
-                "CRYSTALID": "333588",
-                "SCREENCOMPONENTGROUPID": null,
-                "PARENTSAMPLEID": null,
-                "PARENTSAMPLE": null,
-                "BLSUBSAMPLEID": null,
-                "INSPECTIONS": "0",
-                "PROP": "cm14451",
-                "CODE": "",
-                "LOCATION": "169",
-                "ACRONYM": "therm",
-                "PROTEINID": "121393",
-                "SPACEGROUP": "",
-                "COMMENTS": "",
-                "NAME": "H1d1_A1d1_s1",
-                "SHIPMENT": "asdasd",
-                "SHIPPINGID": "7237",
-                "DEWARID": "8584",
-                "DEWAR": "sadsdasd",
-                "CONTAINER": "mycontainer",
-                "CONTAINERID": "34917",
-                "SCLOCATION": null,
-                "SC": "0",
-                "GR": "0",
-                "DC": "0",
-                "AI": "0",
-                "AP": "0",
-                "R": "0",
-                "SCRESOLUTION": null,
-                "SCCOMPLETENESS": null,
-                "DCRESOLUTION": null,
-                "DCCOMPLETENESS": null,
-                "ANOMALOUSSCATTERER": "",
-                "REQUIREDRESOLUTION": null,
-                "CELL_A": null,
-                "CELL_B": null,
-                "CELL_C": null,
-                "CELL_ALPHA": null,
-                "CELL_BETA": null,
-                "CELL_GAMMA": null,
-                "PACKINGFRACTION": null,
-                "DIMENSION1": null,
-                "DIMENSION2": null,
-                "DIMENSION3": null,
-                "SHAPE": null,
-                "THEORETICALDENSITY": null,
-                "CRYSTAL": null,
-                "PROTEIN": "therm",
-                "LOOPTYPE": null,
-                "CENTRINGMETHOD": null,
-                "EXPERIMENTKIND": null,
-                "CONTAINERQUEUEID": null,
-                "QUEUEDTIMESTAMP": null,
-                "COMPONENTNAMES": null,
-                "COMPONENTDENSITIES": null,
-                "COMPONENTIDS": null,
-                "COMPONENTACRONYMS": null,
-                "COMPONENTGLOBALS": null,
-                "COMPONENTAMOUNTS": null,
-                "COMPONENTTYPESYMBOLS": null,
-                "VOLUME": null,
-                "SYMBOL": null,
-                "ABUNDANCE": null,
-                "RECORDTIMESTAMP": "09-10-2020",
-                "RADIATIONSENSITIVITY": null,
-                "ENERGY": null,
-                "USERPATH": null,
-                "SCORE": 0.1,
-                "VALID": 1
-            }
-        ]
-    }}
-  }
-}
-</script>
